requirements.txt
torch>=2.0.0
ultralytics
opencv-python
numpy>=2.2.0,<2.3.0
scipy>=1.15.0
scikit-learn>=1.5.0
pyyaml
tqdm
supervision
filterpy
lap
deep_sort_realtime==1.3.2

